Move update_sa_table to model utils
