+++ mathesar/utils/tables.py
@@ -1,11 +1,6 @@
-from rest_framework.exceptions import ValidationError
-
-from db.tables import (
-    get_table_oids_from_schema, infer_table_column_types, update_table,
-    SUPPORTED_TABLE_UPDATE_ARGS
-)
+from db.tables import get_table_oids_from_schema, infer_table_column_types
 from db.columns import MathesarColumn
-import mathesar.models as models
+from mathesar.models import Table
 
 
 def reflect_tables_from_schema(schema):
@@ -14,10 +9,10 @@ def reflect_tables_from_schema(schema):
         for table in get_table_oids_from_schema(schema.oid, schema._sa_engine)
     }
     tables = [
-        models.Table.objects.get_or_create(oid=oid, schema=schema)
+        Table.objects.get_or_create(oid=oid, schema=schema)
         for oid in db_table_oids
     ]
-    for table in models.Table.objects.all().filter(schema=schema):
+    for table in Table.objects.all().filter(schema=schema):
         if table.oid not in db_table_oids:
             table.delete()
     return tables
@@ -34,10 +29,3 @@ def get_table_column_types(table):
         and not col.foreign_keys
     }
     return col_types
-
-
-def update_sa_table(table, validated_data):
-    for arg in validated_data:
-        if arg not in SUPPORTED_TABLE_UPDATE_ARGS:
-            raise ValidationError({arg: f'Updating {arg} for tables is not supported.'})
-    update_table(table.name, table.schema.name, table.schema._sa_engine, validated_data)
